prepare for division
